代码优化

+++ ymate-platform-persistence-mongodb/src/main/java/net/ymate/platform/persistence/mongodb/impl/MongoGridFsSession.java
@@ -15,6 +15,7 @@
  */
 package net.ymate.platform.persistence.mongodb.impl;
 
+import com.mongodb.client.ClientSession;
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.gridfs.GridFSBucket;
@@ -30,24 +31,21 @@ import net.ymate.platform.core.persistence.impl.DefaultResultSet;
 import net.ymate.platform.persistence.mongodb.IGridFsSession;
 import net.ymate.platform.persistence.mongodb.IMongo;
 import net.ymate.platform.persistence.mongodb.IMongoConnectionHolder;
-import net.ymate.platform.persistence.mongodb.IMongoDataSourceAdapter;
 import net.ymate.platform.persistence.mongodb.support.Operator;
 import net.ymate.platform.persistence.mongodb.support.OrderBy;
 import net.ymate.platform.persistence.mongodb.support.Query;
+import net.ymate.platform.persistence.mongodb.support.QueryBuilder;
+import net.ymate.platform.persistence.mongodb.transaction.Transactions;
 import org.apache.commons.lang.NullArgumentException;
 import org.apache.commons.lang3.StringUtils;
-import org.bson.BsonArray;
 import org.bson.BsonObjectId;
 import org.bson.types.ObjectId;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.InputStream;
+import java.io.*;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.function.Consumer;
-import java.util.stream.Collectors;
 
 /**
  * @author 刘镇 (suninformation@163.com) on 15/11/26 上午11:12
@@ -62,12 +60,15 @@ public class MongoGridFsSession extends AbstractSession<IMongoConnectionHolder> 
 
     private final IMongoConnectionHolder connectionHolder;
 
-    public MongoGridFsSession(IMongoDataSourceAdapter dataSourceAdapter) throws Exception {
-        this(dataSourceAdapter, GridFS.DEFAULT_BUCKET);
+    public MongoGridFsSession(IMongoConnectionHolder connectionHolder) throws Exception {
+        this(connectionHolder, GridFS.DEFAULT_BUCKET);
     }
 
-    public MongoGridFsSession(IMongoDataSourceAdapter dataSourceAdapter, String bucketName) throws Exception {
-        this.connectionHolder = new DefaultMongoConnectionHolder(dataSourceAdapter);
+    public MongoGridFsSession(IMongoConnectionHolder connectionHolder, String bucketName) throws Exception {
+        if (connectionHolder == null) {
+            throw new NullArgumentException("connectionHolder");
+        }
+        this.connectionHolder = connectionHolder;
         this.bucketName = StringUtils.defaultIfBlank(bucketName, GridFS.DEFAULT_BUCKET);
         //
         MongoDatabase mongoDatabase = this.connectionHolder.getConnection();
@@ -96,15 +97,95 @@ public class MongoGridFsSession extends AbstractSession<IMongoConnectionHolder> 
 
     @Override
     public String upload(File file, GridFSUploadOptions fsUploadOption) throws Exception {
+        return upload(null, file, fsUploadOption);
+    }
+
+    @Override
+    public String upload(String id, File file, GridFSUploadOptions fsUploadOption) throws Exception {
+        if (file == null || !file.isAbsolute() || !file.isFile() || !file.exists() || !file.canRead()) {
+            throw new IllegalArgumentException("file");
+        }
         try (InputStream inputStream = new FileInputStream(file)) {
-            return upload(file.getName(), inputStream, fsUploadOption);
+            return upload(id, file.getName(), inputStream, fsUploadOption);
         }
     }
 
     @Override
+    public String upload(File file) throws Exception {
+        return upload(null, file, null);
+    }
+
+    @Override
+    public String upload(String id, File file) throws Exception {
+        return upload(id, file, null);
+    }
+
+    @Override
     public String upload(String fileName, InputStream inputStream, GridFSUploadOptions fsUploadOption) throws Exception {
-        ObjectId returnValue = fsBucket.uploadFromStream(fileName, inputStream, fsUploadOption != null ? fsUploadOption : new GridFSUploadOptions());
+        return upload(null, fileName, inputStream, fsUploadOption);
+    }
+
+    @Override
+    public String upload(String id, String fileName, InputStream inputStream, GridFSUploadOptions fsUploadOption) throws Exception {
+        if (StringUtils.isBlank(fileName)) {
+            throw new NullArgumentException("fileName");
+        }
+        if (inputStream == null) {
+            throw new NullArgumentException("inputStream");
+        }
+        ClientSession clientSession = Transactions.getClientSession(connectionHolder);
+        ObjectId returnValue;
+        if (StringUtils.isNotBlank(id)) {
+            returnValue = new ObjectId(id);
+            if (clientSession == null) {
+                fsBucket.uploadFromStream(new BsonObjectId(returnValue), fileName, inputStream, fsUploadOption != null ? fsUploadOption : new GridFSUploadOptions());
+            } else {
+                fsBucket.uploadFromStream(clientSession, new BsonObjectId(returnValue), fileName, inputStream, fsUploadOption != null ? fsUploadOption : new GridFSUploadOptions());
+            }
+        } else {
+            if (clientSession == null) {
+                returnValue = fsBucket.uploadFromStream(fileName, inputStream, fsUploadOption != null ? fsUploadOption : new GridFSUploadOptions());
+            } else {
+                returnValue = fsBucket.uploadFromStream(clientSession, fileName, inputStream, fsUploadOption != null ? fsUploadOption : new GridFSUploadOptions());
+            }
+        }
         return returnValue.toString();
+    }
+
+    @Override
+    public String upload(String fileName, InputStream inputStream) throws Exception {
+        return upload(null, fileName, inputStream, null);
+    }
+
+    @Override
+    public String upload(String id, String fileName, InputStream inputStream) throws Exception {
+        return upload(id, fileName, inputStream, null);
+    }
+
+    @Override
+    public void download(String id, OutputStream outputStream) throws Exception {
+        if (StringUtils.isBlank(id)) {
+            throw new NullArgumentException("id");
+        }
+        if (outputStream == null) {
+            throw new NullArgumentException("outputStream");
+        }
+        ClientSession clientSession = Transactions.getClientSession(connectionHolder);
+        if (clientSession == null) {
+            fsBucket.downloadToStream(new ObjectId(id), outputStream);
+        } else {
+            fsBucket.downloadToStream(clientSession, new ObjectId(id), outputStream);
+        }
+    }
+
+    @Override
+    public void download(String id, File distFile) throws Exception {
+        if (distFile == null || !distFile.isAbsolute()) {
+            throw new IllegalArgumentException("distFile");
+        }
+        try (OutputStream outputStream = new FileOutputStream(distFile)) {
+            download(id, outputStream);
+        }
     }
 
     @Override
@@ -113,18 +194,38 @@ public class MongoGridFsSession extends AbstractSession<IMongoConnectionHolder> 
     }
 
     @Override
+    public GridFSFile match(String fileHash) {
+        if (StringUtils.isBlank(fileHash)) {
+            throw new NullArgumentException("fileHash");
+        }
+        return findFirst(Query.create(IMongo.GridFs.MD5, Operator.create().eq(fileHash)));
+    }
+
+    @Override
     public GridFSFile findFirst(Query query) {
-        return fsBucket.find(query.toBson()).first();
+        ClientSession clientSession = Transactions.getClientSession(connectionHolder);
+        if (clientSession == null) {
+            return fsBucket.find(query.toBson()).first();
+        }
+        return fsBucket.find(clientSession, query.toBson()).first();
+    }
+
+    @Override
+    public GridFSFile findFirst(QueryBuilder query) {
+        return findFirst(query.build());
     }
 
     @Override
     public GridFSFile find(String id) {
-        return findFirst(Query.create(IMongo.Opt.ID, Operator.create().eq(id)));
+        if (StringUtils.isBlank(id)) {
+            throw new NullArgumentException("id");
+        }
+        return findFirst(Query.create(IMongo.GridFs.ID, Operator.create().eq(new ObjectId(id))));
     }
 
     @Override
     public IResultSet<GridFSFile> find() {
-        return find(null, null, null);
+        return find((Query) null, null, null);
     }
 
     @Override
@@ -134,23 +235,21 @@ public class MongoGridFsSession extends AbstractSession<IMongoConnectionHolder> 
 
     @Override
     public IResultSet<GridFSFile> find(OrderBy orderBy, Page page) {
-        return find(null, orderBy, page);
+        return find((Query) null, orderBy, page);
     }
 
     @Override
     public IResultSet<GridFSFile> find(String filename, OrderBy orderBy) {
+        return find(filename, orderBy, null);
+    }
+
+    @Override
+    public IResultSet<GridFSFile> find(String filename, OrderBy orderBy, Page page) {
         if (StringUtils.isBlank(filename)) {
             throw new NullArgumentException(IMongo.GridFs.FILE_NAME);
         }
         Query query = Query.create(IMongo.GridFs.FILE_NAME, Operator.create().eq(filename));
-        return find(query, orderBy, null);
-//        FindIterable<GridFSFile> findIterable = dbCollection.find(query.toBson());
-//        if (orderBy != null) {
-//            findIterable.sort(orderBy.toBson());
-//        }
-//        List<GridFSFile> results = new ArrayList<>();
-//        findIterable.forEach((Consumer<? super GridFSFile>) results::add);
-//        return new DefaultResultSet<>(results);
+        return find(query, orderBy, page);
     }
 
     @Override
@@ -159,32 +258,28 @@ public class MongoGridFsSession extends AbstractSession<IMongoConnectionHolder> 
     }
 
     @Override
+    public IResultSet<GridFSFile> find(QueryBuilder query) {
+        return find(query.build(), null, null);
+    }
+
+    @Override
     public IResultSet<GridFSFile> find(Query query, OrderBy orderBy) {
         return find(query, orderBy, null);
     }
 
-//    private IResultSet<GridFSFile> doFind() {
-//        FindIterable<GridFSFile> findIterable = dbCollection.find();
-//        if (orderBy != null) {
-//            findIterable.sort(orderBy.toBson());
-//        }
-//        boolean pageFlag = false;
-//        if (page != null && page.page() > 0 && page.pageSize() > 0) {
-//            findIterable.skip((page.page() - 1) * page.pageSize()).limit(page.pageSize());
-//            pageFlag = true;
-//        }
-//        List<GridFSFile> results = new ArrayList<>();
-//        findIterable.forEach((Consumer<? super GridFSFile>) results::add);
-//        return pageFlag ? new DefaultResultSet<>(results, page.page(), page.pageSize(), dbCollection.countDocuments()) : new DefaultResultSet<>(results);
-//    }
+    @Override
+    public IResultSet<GridFSFile> find(QueryBuilder query, OrderBy orderBy) {
+        return find(query.build(), orderBy, null);
+    }
 
     @Override
     public IResultSet<GridFSFile> find(Query query, OrderBy orderBy, Page page) {
+        ClientSession clientSession = Transactions.getClientSession(connectionHolder);
         GridFSFindIterable findIterable;
         if (query != null) {
-            findIterable = fsBucket.find(query.toBson());
+            findIterable = clientSession == null ? fsBucket.find(query.toBson()) : fsBucket.find(clientSession, query.toBson());
         } else {
-            findIterable = fsBucket.find();
+            findIterable = clientSession == null ? fsBucket.find() : fsBucket.find(clientSession);
         }
         if (orderBy != null) {
             findIterable.sort(orderBy.toBson());
@@ -193,11 +288,10 @@ public class MongoGridFsSession extends AbstractSession<IMongoConnectionHolder> 
         boolean pageFlag = false;
         if (page != null && page.page() > 0 && page.pageSize() > 0) {
             findIterable.skip((page.page() - 1) * page.pageSize()).limit(page.pageSize());
-            //
             if (query != null) {
-                recordCount = dbCollection.countDocuments(query.toBson());
+                recordCount = clientSession == null ? dbCollection.countDocuments(query.toBson()) : dbCollection.countDocuments(clientSession, query.toBson());
             } else {
-                recordCount = dbCollection.countDocuments();
+                recordCount = clientSession == null ? dbCollection.countDocuments() : dbCollection.countDocuments(clientSession);
             }
             pageFlag = true;
         }
@@ -207,8 +301,18 @@ public class MongoGridFsSession extends AbstractSession<IMongoConnectionHolder> 
     }
 
     @Override
+    public IResultSet<GridFSFile> find(QueryBuilder query, OrderBy orderBy, Page page) {
+        return find(query.build(), orderBy, page);
+    }
+
+    @Override
     public IResultSet<GridFSFile> find(Query query, Page page) {
         return find(query, null, page);
+    }
+
+    @Override
+    public IResultSet<GridFSFile> find(QueryBuilder query, Page page) {
+        return find(query.build(), null, page);
     }
 
     @Override
@@ -219,19 +323,31 @@ public class MongoGridFsSession extends AbstractSession<IMongoConnectionHolder> 
         if (StringUtils.isBlank(newFileName)) {
             throw new NullArgumentException("newFileName");
         }
-        fsBucket.rename(new ObjectId(id), newFileName);
+        ClientSession clientSession = Transactions.getClientSession(connectionHolder);
+        if (clientSession == null) {
+            fsBucket.rename(new ObjectId(id), newFileName);
+        } else {
+            fsBucket.rename(clientSession, new ObjectId(id), newFileName);
+        }
     }
 
     @Override
     public void remove(String id) {
-        fsBucket.delete(new ObjectId(id));
+        if (StringUtils.isBlank(id)) {
+            throw new NullArgumentException("id");
+        }
+        ClientSession clientSession = Transactions.getClientSession(connectionHolder);
+        if (clientSession == null) {
+            fsBucket.delete(new ObjectId(id));
+        } else {
+            fsBucket.delete(clientSession, new ObjectId(id));
+        }
     }
 
     @Override
     public void remove(Collection<String> ids) {
         if (ids != null && !ids.isEmpty()) {
-            BsonArray bsonIds = ids.stream().map(id -> new BsonObjectId(new ObjectId(id))).collect(Collectors.toCollection(BsonArray::new));
-            fsBucket.delete(bsonIds);
+            ids.stream().filter(StringUtils::isNotBlank).map(ObjectId::new).forEach(fsBucket::delete);
         }
     }
 }
